requirements.txt
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
ffmpeg==1.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
magic-filter==1.0.12
multidict==6.0.4
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
psutil==5.9.7
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
pytz==2023.3.post1
sniffio==1.3.0
typing_extensions==4.9.0
yarl==1.9.4
